Create FlatFeedApi to extend what a FlatFeed can do
